backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6
google-genai>=1.16.0
python-dotenv==1.0.0
faiss-cpu==1.9.0.post1
langchain-text-splitters==0.3.11
numpy==1.26.4